requirements.txt
requests>=2.28.0
pycryptodome>=3.18.0
tqdm>=4.65.0
